Logging: Fix conflicting type for smartTime()

+++ src/utils/logging.h
@@ -50,6 +50,6 @@ C-ray internal formatted logger.
 */
 void logr(enum logType type, const char *fmt, ...);
 
-void smartTime(unsigned long long milliseconds, char *buf);
+void smartTime(uint64_t milliseconds, char *buf);
 
 void printSmartTime(unsigned long long ms);
